Overhaul and fix resolution of dependencies
This changes how dependencies get resolved. With this change, package dependencies
are resolved using both the package's declared dependencies, as well as the originating
project's dependencies.

This removes the previous approach of placing dependencies in an `editor-support` directory.

The resolved types of all package dependencies are resolved according to the enclosing project's
declared dependencies.